+++ physicsnemo/cfd/evaluation/inference/wrappers/surface_baseline.py
@@ -48,6 +48,7 @@ class SurfaceBaselineWrapper(CFDModel):
 
     INFERENCE_DOMAIN: ClassVar[InferenceDomain] = "surface"
     OUTPUT_LOCATION: ClassVar[OutputLocation] = "cell"
+    REQUIRES_REMOTE_ASSETS: ClassVar[bool] = False
 
     @property
     def output_location(self) -> OutputLocation: